requirements.txt
langchain==0.1.20
langchain-openai==0.1.7
langchain-community==0.0.38
langgraph==0.0.40
yfinance>=0.2.40
requests>=2.31.0
pandas>=2.1.0
numpy>=1.24.0
beautifulsoup4>=4.12.0
feedparser>=6.0.10
newspaper3k>=0.2.8
python-dotenv>=1.0.0
tavily-python>=0.3.0
duckduckgo-search>=4.1.0
streamlit>=1.36.0
plotly>=5.22.0
fastapi>=0.111.0
uvicorn>=0.30.0
